requirements.txt
amqp==5.3.1
APScheduler==3.6.3
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
bcc==0.1.10
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==4.2.2
celery==5.3.6
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
idna==3.10
kombu==5.5.4
numpy==2.2.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.0.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
traitlets==5.14.3
traittypes==0.2.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
